Logical operators

Test an object x. See Details for an explanation of how each function operates.

is_char_numeric(x)

is_formula(x)

Arguments

x

an object to check.

Value

logical

Details

is_char_numeric

TRUE if x can be converted into a numeric vector by as.numeric without coercing character elements without a commonly-understood numeric representation.

is_formula

TRUE if x is a formula object.

Examples

x <- c(1, 3, "3", NA)
xx <- c(x, "A")
is_char_numeric(x)

#> [1] TRUE

f <- "y ~ x"
is_formula(f)

#> [1] FALSE

ff <- as.formula(f)
is_formula(ff)

#> [1] TRUE

## Not run: ------------------------------------
# is_char_numeric(xx) # FALSE
## ---------------------------------------------