Wrap a string using newline delimeters

Wrap a string using newline delimeters

text_wrap(x, n = 20L)

Arguments

x

a character vector

n

an integer of length 1. text_wrap() will silently coerce any numeric value passed to n to an integer.

Value

a version of x in which each element substitutes newline delimiters for the immediate next space (e.g. ' ') after every nth character.

References

I attribute the regular expression that powers text_wrap() to StackOverflow user xiechao: http://stackoverflow.com/a/2352006/3277821

Examples

s <- "The quick brown fox jumped over the lazy dog."
text_wrap(s)

#> [1] "The quick brown fox\njumped over the lazy\ndog."
#> 

text_wrap(s, 10)

#> [1] "The quick\nbrown fox\njumped\nover the\nlazy dog."
#> 

cat(paste0(text_wrap(s), "\n"))

#> The quick brown fox
#> jumped over the lazy
#> dog.
#>