Remove trailing and leading spaces from a character vector

Remove trailing and leading spaces from a character vector

trim(x)

Arguments

x

a character vector

Value

a version of x without trailing OR leading spaces

See also

str_trim

Examples

trim("foo ")

#> [1] "foo"
#> 

trim(" bar ")

#> [1] "bar"
#> 

trim("baz")

#> [1] "baz"
#>